Update 'latest' workflows to enable cross-repo testing (#940)
This changes the workflow to publish charts with versions that are unique rather than using build metadata.

Instead of `VERSION+COMMIT` we now have `VERSION-COMMIT`. This will cause all of the chart versions _not_ to be considered equal, as was the problem before. I have not yet re-enabled the workflow on-push, but I will after we have confirmed that everything works end-to-end.

Update: I've also now added a build-time environment variabel, `FLUVIO_VERSION_SUFFIX`, which should be set to contain the git hash of the checked-out branch. When it is present, the VERSION that gets baked into the cluster and CLI will be something like `0.7.4-alpha.0-abcdef`, where `abcdef` is the FLUVIO_VERSION_SUFFIX` that was passed. This variable will only be set in the "fluvio_latest" workflows, in regular builds and releases it will be absent.