Better, more automated release flow (#1010)
This is not quite done yet, but I wanted to open a PR so we can start reviewing it since I'm sure there will be questions and things to fix. This introduces a new workflow for CI/CD that makes the deployment of "latest" builds faster and more robust. Let me explain the new workflow:

We have two primary workflow files, `ci.yml` and `publish_latest.yml`.

The CI workflow now has two distinct "modes":

- One is for running checks and tests on PRs, this mode runs every time a commit is pushed to a PR and it builds and runs tests in debug mode.
- The other mode gets triggered whenever a commit is pushed to `staging`, i.e. whenever Bors is executing an `r+` build. In this mode, everything is built in release mode, and the `fluvio` and `fluvio-run` artifacts for each target are saved using `actions/upload-artifact@v2`. This means that if Bors pushes a commit to master, we are guaranteed to have good release artifacts already built and saved.

The `publish_latest` workflow is triggered any time a commit is pushed to `master`, i.e. whenever Bors successfully runs CI and build artifacts are saved. This workflow will download the artifacts that were saved by CI and do the following:

- Publish the artifacts to the `dev` GitHub prerelease
- Build a docker image wrapping `fluvio-run` on linux-musl and push to docker registry
- Push the latest helm chart
- Publish the CLI to packages.fluvio.io and bump the latest CLI version.

The final piece to this work is finishing the `release` workflow, which should essentially just re-tag everything from the most recent "latest" publishing. The objective is that this workflow is very minimal and has very little that can go wrong. Previously, our release workflow would actually try to build the release artifacts, making it very slow as well as susceptible to tripping over flaky builds.

---

There is one foreseen obstacle remaining for finishing the release workflow. When building our components, we bake in the current value of the `VERSION` file. When building "latest" versions, we bake in the git commit into the version as well. This was done in order to satisfy a problem with the platform_version check between a client and cluster, but I now believe that was the wrong choice of approach.

The problem this creates for us now is that when we build release artifacts, they will have git commits baked into the VERSION as a prerelease tag of the version. This is something we do not want when we publish a proper release, so this would prevent us from re-using build artifacts saved from a "latest" workflow as a proper release. I have a fix in mind for this, I just need to do it now.