First pass at making release workflow idempotent (#1058)
We now check for the existence of artifacts before doing any pushing or retagging.

Also, we can give the action a git commit which it will use as a base (with respect to the `master` branch). This will allow us to re-run the release job in the event of environmental failures.

We also fix the helm chart naming, which used to include git hash, but should not during Release. 

Resolves #1056

Example runs:
Fail, on account of being unable to push helm chart - https://github.com/infinyon/fluvio-gh-actions-dev/actions/runs/814121338

This is an override to the helm chart check, just to show off the bump stable version step - Don't worry about the bump stable fluvio job not depending on release helm job. I needed to test the idempotency of the bump job.
https://github.com/infinyon/fluvio-gh-actions-dev/actions/runs/813947722