Add OpenTelemetry/Jaeger tracing backend (#1118)
This is the first step towards improving our observability within Fluvio. This adds a jaeger backend to our tracing setup, so you can now run a jaeger instance and collect traces.

- Adds Jaeger backend to tracing subscriber for SC and SPU
- Adds more tracing instrumentation to the SPU

To try this out for yourself, run Jaeger in a docker container with the following command:

```
docker run -d -p6831:6831/udp -p6832:6832/udp -p16686:16686 -p14268:14268 jaegertracing/all-in-one:latest
```

Then open up `http://localhost:16686` in your browser. Try running some Fluvio commands to interact with the cluster, then view traces sorted by `fluvio-spu` or `fluvio-sc`.

I don't know what we want the ultimate scope of this PR to be. It could potentially be merged now, but do we want to explore any of the following first?

- Add more instrumentation to SC? I think that'd be a good idea
- Add instrumentation to the client, i.e. Fluvio CLI? If yes, we would want to make sure it is optional and probably disabled by default
- Try adding prometheus metrics? There are examples showing how this works, but I believe it requires creating a metrics route on an HTTP server, so it would really only make sense for the SC and SPU. This is probably out of scope for this.