Give user feedback when consuming in interactive mode (#1171)
Closes #700 

When using the Consumer CLI in "interactive mode", the CLI will now print a little message about how it is going to consume.

<img width="669" alt="image" src="https://user-images.githubusercontent.com/4210949/122231966-e0f5a080-ce88-11eb-8073-d11d3ca12848.png">

The most important one here is in the case that no flags are given, e.g. `fluvio consume one`. This was confusing to the user in the past because it would appear to just hang as if it were loading, when it was actually consuming from the end of the partition and just waiting for records to arrive. Now there's a message to say that.