Clean up stream_fetch code (#1188)
This is just a small cleanup task as I am working on aggregate SmartStreams. Right now all of the SmartStream engine stuff is in the `filter.rs` module in the SPU, but if we are going to be reusing it (e.g. in `aggregate.rs`) then we should move this stuff up one level to be more "common" code.

There should not be any changes in here except for moving the location of this code.