Update cache keys to prevent conflicting saves (#1265)
In our CI, our slowest job is "Unit/Doc test", and it appears that our rust-cache setup is failing to cache for that job:

![image](https://user-images.githubusercontent.com/4210949/126163475-1bd4d772-97de-48dd-8927-38eb49948164.png)

I believe this is because this job is part of a matrix and only the first job in that matrix is "reserving" the key to cache with. This PR should give each job in the matrix a unique key so that caching can work properly.