Fix plugin install (#1434)
Adds a new error variant representing "this package isn't available" for a given package, version, and target. Also catches that error when running `fluvio install` in order to print a nice message to the user instead of crashing. Now, if somebody uses the installer from a platform that is not supported (or they try to specify a VERSION that does not exist), they'll get a message like the following in their installer output:

<img width="672" alt="image" src="https://user-images.githubusercontent.com/4210949/129376758-cc9e0a5d-3667-4416-a33e-fb13f7e06aac.png">