Create SmartStreamBase to manage wasmtime/memory (#1552)
This pulls a lot of the `wasmtime` code related to reading and writing memory into a common type called `SmartStreamBase`. Each concrete smartstream wraps this as well as their own typed wasm function and potentially any state (e.g. the aggregate accumulator).

This is part 1 of the smartstreams refactor in preparation for #1335