Added spu local address logic (#1591)
Closes #1537.
Closes  #1546.

Tried to make this a non-breaking change. Looks like with the optional type for `public_endpoint_local`, it works as expected.

Testing this was a bit of a pain, we ended up spinning up a new pod in the k8s cluster and running the fluvio client with `use_spu_local_address = true` in the config to verify that it was using the correct endpoint.