Support `ReadCommitted` isolation in SPU for Produce requests (#2336)
The produce handler uses the `acks` field from `ProduceRequest` to determine which isolation will be taken. 
In case of `ReadCommitted` isolation, it waits until the replica's `hw` includes records from the requests before returning the response. Awaiting is limited by the `timeout_ms` field of the request. For now, the timeout is not shared between partitions, it is counted separately for each partition. 

For `ReadUncommitted` there is no waiting.

Related #2302