Fix `wasi` functions binding relying on order (#2428)
The binding of host function `copy_record` for `wasi` expects that the function always goes first in the list of imports. This is not always correct. Instead of lookup by order, we will do a lookup by name.
Also, the PR adds integration tests that will check the smart module instantiation works for both `wasm` and `wasi` cases.