fix: Add non-interactive `smdk generate` flow (#2693)
Adding structure for customizing the results of `smdk generate` through CLI flags as an alternative to the default user prompts.


CLI tests run `smdk generate` to build each type of SmartModule with all combinations of opt-in option in `cargo-generate.toml` file. Then we  build it with `smdk build`. These cover both stable and latest builds.
(however most are being skipped at the moment, in preparation for after release)

Opinionated changes:
- Add SmartModule params visibility from the templates
- Provide developer focused options in CLI hidden under long help `--help` (short `-h` hides them)
  - CLI descriptions set expectations that they are not recommended for use.
  - `--develop` flag for testing with latest `fluvio-smartmodule`
- Renamed  structs (more verbose) for clarity between cargo templates sources, and the template variables.
- Removed the `smartmodule-init` variable and just using `smartmodule-params`
  - This was done after removing the existing form of params, since it is all moving into init