feat(socket): moved versioned socket to fluvio-socket crate (#2797)
`VersionedSocket` and `VersionedSerialSocket` will be needed for use by crates other than `fluvio`. The first will be the SmartModule API client that will communicate with SC and can be used by other components. Therefore, we need versioned sockets in the more specialized crate to avoid bringing a dependency to `fluvio` crate.