fix: disable cargo vet (#2807)
Disable `cargo vet` until the database for audition is ready. Currently
our audition database is not importing any audition files from other
trusted resources, and dependencies out of the range of these databases
must be allowed/exempted by ourselves.

Until we have defined a set of supply chains to trust for import and also
established which deps to exempt, we must disable to allow incoming PRs.