cli: Add `fluvio cluster status` (#2824)
### Behavior

Adds `fluvio cluster status` to the CLI

Tries to connect to the SC, reports "none" if unsuccessful, describes where (local/k8s/cloud) cluster is deployed if successful. If the SC is up, it then will report if the SPUs are down, or if they are all empty.

### Comment

I'm not sure if this was still wanted but it looked like a fun first issue. It's also possible I misinterpreted the original request. Specifically, I'm not sure how we could determine if the status should be "stopped (no sc, spu running but data exits)", as I'm not sure how to determine if there is data in the spus if the connection to the sc fails.

I would love some feedback and would be happy to make modifications!

https://github.com/infinyon/fluvio/issues/1673