feat: separate custom config from common config (#2917)
This PR changes the shape of the connector's config separating fields for custom user-defined config from `ConnectorConfig`. For example:
```yaml
meta:
  version: 0.1.0
  name: my-json-test-connector
  type: json-test-source
  topic: test-topic
json:
  interval: 10  
  template: '{"template":"test"}'  
transforms:
  - uses: infinyon/jolt@0.1.0
    with:
```
sections `meta` and `transforms` go to `ConnectorConfig`, `json` goes to custom config (the name is configurable, by default  is `custom`). If the name of the custom config is overridden, need to specify it in the package metadata also.