Update segment.rs (#2950)
Currently when using `Isolation::ReadCommitted`, we read records until current HW. https://github.com/infinyon/fluvio/blob/934544dec2403cd65113eda5c2a6fb6b509793d4/crates/fluvio-storage/src/replica.rs#L106

Here, there are two scenarios. If Start offset is greater or equal than base offset of current active segment. Then read_records is call in that segment. If HW is greater than last offset in that segment it will fail. (I think that this scenario should not happen since current active segment should be ok)
https://github.com/infinyon/fluvio/blob/934544dec2403cd65113eda5c2a6fb6b509793d4/crates/fluvio-storage/src/replica.rs#L348

On the other hand, if the start offset is in a old segment, it calls find_slice, that first checks in which segment is the start offset and then call `read_records` there. This read_records will fail if that HW offset is not in the current segment.

https://github.com/infinyon/fluvio/blob/934544dec2403cd65113eda5c2a6fb6b509793d4/crates/fluvio-storage/src/replica.rs#L361

This PR updates read_records to truncate max_offset to end_offset of current segment if it is larger than that so we can read records of that segment without failures


Co-authored-by: Luis Moreno <morenol@users.noreply.github.com>