CI: Optimize smdk Smoke Test (#2972)
https://github.com/infinyon/fluvio/issues/2925

The majority of the run time was spent repeatedly building the same dependencies for each Smart Module. This solution generates each Smart Module within a Cargo workspace, which allows them to share dependencies after building once.

An _alternative solution_ would be to develop a scheme for Github Workflows to cache the dependencies for each Smart Module. The Github Workflows caching solution may be more conventional but has its drawbacks:

- would couple the test suite to Github Workflows
- test would still run slow in local dev environments
- test would still take 10-20 minutes on cache miss