fix: relax `cdk` secrets validation (#3093)
Currently, in `cdk deploy`, we validate strict correspondence between metadata of secrets defined in `Connector.toml` and secrets presented in the config file. This is incorrect because, in the config file, secrets might not be used or used as a raw string(`SecretString` supports using secrets by its name or raw string value).

Fixed by removing this part of validation. 

Fixes #3087