feat: explicit set of errors to retry (#3137)
The current default time to wait for Fluvio Socket is 5 minutes, and until that time, if there is a network connectivity issue, it does not report an error. It seems too coarse, it doesn't align well with retries. For the producer, for example, we set a whole timeout for all retries is 5 minutes.

This PR is a proposal to decrease this value and explicitly define the set of socket errors that we want to retry.

Related #3124