feat: introduce utilily to check k8s resource names (#3163)
Provides `shared::is_valid_k8s_resource_name` which checks if the provided
string is a k8s resource name compliant value. This could be used for topics,
connectors and other resources relying on k8s. In the future is likely we want to
use a single function all around to apply the same rules on a centralized approach.