feat: Secrets in connector meta (#3194)
Previously, we defined possible secrets used by a connector on connector package spec. This approach makes less customizable the experience for the user.

In this PR we are moving this definition to the connector meta configuration, users can define explicitly the name of the secrets that they want and use them explicitly.

This change also implies that secret names are not longer detected from parsing the connector specific configuration and instead only parsing the secrets section on the connector configuration.