fix: fix fluvio hub utils (#3198)
In the update of `pem` https://github.com/infinyon/fluvio/commit/fdcfce51067a44c06a91bc8e4aab518f0a193145 to `2.0.0`. It was changed the tag from `PUBLIC KEY` and `PRIVATE KEY` to `PUBLIC_KEY` and `PRIVATE_KEY`. This changes uses a global const so we use the same tag when we write to file and when we read from file