refactor: simplified control plane api (#3073)
This PR implements a more generalized admin API.

The existing Admin API uses a macro to map the admin type to a global object type.  While this works, it requires maintaining a mapping relationship every time a new admin type is added.   In addition, mapping for each action (create, delete, watch) has a slightly different kind of encoding, which adds to the complexity.   The underlying complexity is that with the current Rust typing system, converting Traits with associated properties to trait objects is not allowed.  Hence use of macro to map handle encoding and decoding.

With this PR, type mapping is lazily done.  Borrowing from `Arc` and `Anyhow` approaches, downcasting/upcasting is delayed much as possible.  This preserves type-specific Admin type calls, but extending to any arbitrary admin object types is possible.

The core of the new approach is this trait:
```
pub trait TryEncodableFrom<T>: Sized + Encoder + Decoder {
        fn try_encode_from(value: T, version: Version) -> Result<Self>;

        fn downcast(&self) -> Result<Option<T>>;
    }
```

Using fluvio's encoder/decoder, any arbitrary can be mapped into the binary blog and vice versa.   Once received from the network, the encoded type can be downcast to a specific type as necessary.

The changes are fully backward compatible.  The existing macro is used when talking to the previous version of Fluvio.   The backward compatibility codes (classic) will be removed in the future.