feat/move-specs-traits-to-fluvio (#3598)
This is preparation for implementing a local metadata store.

Added a new `MetadataClient` interface that is owned by fluvio, and the metadata engine switched to it. This trait describes the contract that is needed for any metadata storage not only k8s.

All types that implement `MetadataClient` from `k8-metadata-client` also implement the new interface.

Also, did some cleanups to reduce the amount of generics in the API for better readability.